Move errmatch to internal/mirror and align file naming
- Error matchers are only used by mirror commands, belong in internal
- File names match function names: classify.go -> diagnose.go (better name for clarity in general)
